Refactor and reorganize GUI namespaces and structure.
Moved classes into more appropriate namespaces for clarity and modularity. Updated font initialization logic and added dock space functionality. Cleaned up unused code and streamlined imports.
