+++ Gui/Rendering/Viewport.cs
@@ -1,6 +1,6 @@
 ﻿using System.Numerics;
 
-namespace Tyr.Gui;
+namespace Tyr.Gui.Rendering;
 
 public record Viewport(
     Vector2 Offset, // Top-left in screen space (e.g. from GetCursorScreenPos)
